Secured inputs passed on login page
