+++ app/Http/Controllers/RegisterController.php
@@ -61,7 +61,7 @@ class RegisterController extends Controller {
             
             $credentials = $request->only('email', 'password');
             if (Auth::attempt($credentials)) {
-                return redirect()->route('inventory');
+                return redirect()->route('home');
             }
         }
         return register();
