+++ resources/views/ForgotPassword/forgotPassword.blade.php
@@ -29,5 +29,4 @@
         </form>
     </div>
 </div>
-
 @endsection
